requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic_settings
pymysql[rsa]
cryptography
bcrypt==4.1.2
